Relocate xlsx_helper.
