--- docs/api_reference/trestle.common.xlsx_helper.md
@@ -1,2 +1,0 @@
-::: trestle.common.xlsx_helper
-handler: python
