+++ docs/api_reference/trestle.tasks.xlsx_helper.md
@@ -1,0 +1,2 @@
+::: trestle.tasks.xlsx_helper
+handler: python
